// src/generated/interface/schemas/graphviz_high_level/unconstrained.ts
export type _T_Graph = {
    readonly 'name': _pt.Optional_Value<string>
    readonly 'tree': _T_Tree
    readonly 'type': _i_core._T_State_Group<null, 
        | readonly ['directed', {
            readonly 'edges': _i_core._T_List<null, {
                readonly 'attributes': _T_Edge_Attributes
                readonly 'from': _T_End_Point_Specification
                readonly 'to': _T_End_Point_Specification
            }>
        }]
        | readonly ['undirected', {
            readonly 'edges': _i_core._T_List<null, {
                readonly 'attributes': _T_Edge_Attributes
                readonly 'yang': _T_End_Point_Specification
                readonly 'yin': _T_End_Point_Specification
            }>
        }]
    >
}
// ...
export type _T_Tree = {
    readonly 'attributes': _T_Graph_Attributes
    readonly 'elements': _i_core._T_Dictionary<null, _i_core._T_State_Group<null, 
        | readonly ['node', {
            readonly 'attributes': _T_Node_Attributes
        }]
        | readonly ['sub', {
            readonly 'tree': _T_Tree
            readonly 'type': _i_core._T_State_Group<null, 
                | readonly ['cluster', null]
                | readonly ['group', null]
                | readonly ['subgraph', null]
            >
        }]
    >>
}

// **** FRIENDLY NAMES FOR THE GLOBAL TYPES

export type Attribute_Value = _T_Attribute_Value

export type Edge_Attributes = _T_Edge_Attributes

export type End_Point_Specification = _T_End_Point_Specification

export type Graph = _T_Graph

export type Graph_Attributes = _T_Graph_Attributes

export type Node_Attributes = _T_Node_Attributes

export type Tree = _T_Tree
// ...
    export namespace SG {
        export type html_string = string
        export type _number = string
        export type _string = string
    }
    export type SG = 
        | readonly ['html string', string]
        | readonly ['number', string]
        | readonly ['string', string]
// ...
        export namespace O {
            
            export namespace compass_direction {
                export type O = string
            }
            export type compass_direction = _pt.Optional_Value<string>
            export type port = string
        }
// ...
    export namespace name {
        export type O = string
    }
    export type name = _pt.Optional_Value<string>
    
    export namespace tree {
    }
    export type tree = _T_Tree
// ...
                        export namespace attributes {
                        }
                        export type attributes = _T_Edge_Attributes
                        
                        export namespace _from {
                        }
                        export type _from = _T_End_Point_Specification
                        
                        export namespace to {
                        }
                        export type to = _T_End_Point_Specification
// ...
                export type edges = _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'from': _T_End_Point_Specification
                    readonly 'to': _T_End_Point_Specification
                }>
// ...
            export type directed = {
                readonly 'edges': _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'from': _T_End_Point_Specification
                    readonly 'to': _T_End_Point_Specification
                }>
            }
// ...
    export type _type = _i_core._T_State_Group<null, 
        | readonly ['directed', {
            readonly 'edges': _i_core._T_List<null, {
                readonly 'attributes': _T_Edge_Attributes
                readonly 'from': _T_End_Point_Specification
                readonly 'to': _T_End_Point_Specification
            }>
        }]
        | readonly ['undirected', {
            readonly 'edges': _i_core._T_List<null, {
                readonly 'attributes': _T_Edge_Attributes
                readonly 'yang': _T_End_Point_Specification
                readonly 'yin': _T_End_Point_Specification
            }>
        }]
    >
// ...
export namespace Graph {
    
    export namespace name {
        export type O = string
    }
    export type name = _pt.Optional_Value<string>
    
    export namespace tree {
    }
    export type tree = _T_Tree
    
    export namespace _type {
        
        export namespace SG {
            
            export namespace directed {
                
                export namespace edges {
                    
                    export namespace L {
                        
                        export namespace attributes {
                        }
                        export type attributes = _T_Edge_Attributes
                        
                        export namespace _from {
                        }
                        export type _from = _T_End_Point_Specification
                        
                        export namespace to {
                        }
                        export type to = _T_End_Point_Specification
                    }
                    export type L = {
                        readonly 'attributes': _T_Edge_Attributes
                        readonly 'from': _T_End_Point_Specification
                        readonly 'to': _T_End_Point_Specification
                    }
                }
                export type edges = _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'from': _T_End_Point_Specification
                    readonly 'to': _T_End_Point_Specification
                }>
            }
            export type directed = {
                readonly 'edges': _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'from': _T_End_Point_Specification
                    readonly 'to': _T_End_Point_Specification
                }>
            }
            
            export namespace undirected {
                
                export namespace edges {
                    
                    export namespace L {
                        
                        export namespace attributes {
                        }
                        export type attributes = _T_Edge_Attributes
                        
                        export namespace yang {
                        }
                        export type yang = _T_End_Point_Specification
                        
                        export namespace yin {
                        }
                        export type yin = _T_End_Point_Specification
                    }
                    export type L = {
                        readonly 'attributes': _T_Edge_Attributes
                        readonly 'yang': _T_End_Point_Specification
                        readonly 'yin': _T_End_Point_Specification
                    }
                }
                export type edges = _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'yang': _T_End_Point_Specification
                    readonly 'yin': _T_End_Point_Specification
                }>
            }
            export type undirected = {
                readonly 'edges': _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'yang': _T_End_Point_Specification
                    readonly 'yin': _T_End_Point_Specification
                }>
            }
        }
        export type SG = 
            | readonly ['directed', {
                readonly 'edges': _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'from': _T_End_Point_Specification
                    readonly 'to': _T_End_Point_Specification
                }>
            }]
            | readonly ['undirected', {
                readonly 'edges': _i_core._T_List<null, {
                    readonly 'attributes': _T_Edge_Attributes
                    readonly 'yang': _T_End_Point_Specification
                    readonly 'yin': _T_End_Point_Specification
                }>
            }]
    }
    export type _type = _i_core._T_State_Group<null, 
        | readonly ['directed', {
            readonly 'edges': _i_core._T_List<null, {
                readonly 'attributes': _T_Edge_Attributes
                readonly 'from': _T_End_Point_Specification
                readonly 'to': _T_End_Point_Specification
            }>
        }]
        | readonly ['undirected', {
            readonly 'edges': _i_core._T_List<null, {
                readonly 'attributes': _T_Edge_Attributes
                readonly 'yang': _T_End_Point_Specification
                readonly 'yin': _T_End_Point_Specification
            }>
        }]
    >
}
// ...
export namespace Tree {
    
    export namespace attributes {
    }
    export type attributes = _T_Graph_Attributes
    
    export namespace elements {
        
        export namespace D {
            
            export namespace SG {
                
                export namespace node {
                    
                    export namespace attributes {
                    }
                    export type attributes = _T_Node_Attributes
                }
                export type node = {
                    readonly 'attributes': _T_Node_Attributes
                }
                
                export namespace sub {
                    
                    export namespace tree {
                    }
                    export type tree = _T_Tree
                    
                    export namespace _type {
                        
                        export namespace SG {
                            export type cluster = null
                            export type group = null
                            export type subgraph = null
                        }
                        export type SG = 
                            | readonly ['cluster', null]
                            | readonly ['group', null]
                            | readonly ['subgraph', null]
                    }
                    export type _type = _i_core._T_State_Group<null, 
                        | readonly ['cluster', null]
                        | readonly ['group', null]
                        | readonly ['subgraph', null]
                    >
                }
                export type sub = {
                    readonly 'tree': _T_Tree
                    readonly 'type': _i_core._T_State_Group<null, 
                        | readonly ['cluster', null]
                        | readonly ['group', null]
                        | readonly ['subgraph', null]
                    >
                }
            }
            export type SG = 
                | readonly ['node', {
                    readonly 'attributes': _T_Node_Attributes
                }]
                | readonly ['sub', {
                    readonly 'tree': _T_Tree
                    readonly 'type': _i_core._T_State_Group<null, 
                        | readonly ['cluster', null]
                        | readonly ['group', null]
                        | readonly ['subgraph', null]
                    >
                }]
        }
        export type D = _i_core._T_State_Group<null, 
            | readonly ['node', {
                readonly 'attributes': _T_Node_Attributes
            }]
            | readonly ['sub', {
                readonly 'tree': _T_Tree
                readonly 'type': _i_core._T_State_Group<null, 
                    | readonly ['cluster', null]
                    | readonly ['group', null]
                    | readonly ['subgraph', null]
                >
            }]
        >
    }
    export type elements = _i_core._T_Dictionary<null, _i_core._T_State_Group<null, 
        | readonly ['node', {
            readonly 'attributes': _T_Node_Attributes
        }]
        | readonly ['sub', {
            readonly 'tree': _T_Tree
            readonly 'type': _i_core._T_State_Group<null, 
                | readonly ['cluster', null]
                | readonly ['group', null]
                | readonly ['subgraph', null]
            >
        }]
    >>
}